Added basic FastAPI implementation
Added base routes for the application,

`/` for the base homepage for Novel Crawlers.
`/add-book` to add the information for a book.
`/books/{book_name}` for the different books which provide information
regarding the book.
`/not-found` as a redirect for incomplete routes/functionality in the
application.

`/docs` and `/redoc` can be used as a means to explore the different
routes and the API specification.

The changes that are yet to be made,
- [ ] Addition of a UI.
- [ ] Addition of a Form.
- [ ] Connecting it to a database.
- [ ] Addition of links to the various different routes.