Removed user, working model created.
The `user` might have been a cause for a permission issue during
development, and therefore has been discarded in favor of getting this
thing to work.

There is now a working model for the novel-crawlers, and while the
option to provide a directory for the output is present, the line in the
`docker-compose.yml` is to be altered to set the path for the book.

This is of the form `/path/to/files/TWI:/data/files`, so, when sending
in a request for the script to begin it's execution, we will always be
required to specify `/data/files`.

There is an idea to just hardcode the path location, when sending in the
request, which is very feasible.

However, this is now a working model, so, yay!